chore: move enum type class + formatting changes
