+++ src/main/java/com/data/dataxer/models/dto/DocumentPackDTO.java
@@ -1,6 +1,6 @@
 package com.data.dataxer.models.dto;
 
-import com.data.dataxer.Enums.DocumentType;
+import com.data.dataxer.models.enums.DocumentType;
 import lombok.Getter;
 import lombok.Setter;
